Improving module organisation for Creator classes
